Change colors to theme
