+++ packages/web/src/Base/styles/index.js
@@ -1,7 +1,3 @@
-import colors from './colors';
 import theme from './theme';
 
-export default {
-  colors,
-  theme
-}
+export default theme;
